refactor(tonk-schema): split analyzer.rs into focused submodules
The 2340-line `analyzer.rs` is split across `analyzer/` per
concern:

- `analyzer/error.rs` — `AnalyzeError`
- `analyzer/resolver.rs` — `Resolver` trait + `Resolved*`,
  `NoopResolver`, `ResolverError`
- `analyzer/scope.rs` — in-doc name index (`Scope`), parking_lot
  guards
- `analyzer/declaration.rs` — `attribute!` / `concept!` body
  parsing + their `Application` builders + `attribute_schema` /
  `concept_schema` helpers
- `analyzer/query.rs` — `build_query_application` + its
  `this`-slot helper
- `analyzer/assertion.rs` — `build_assertion_application`,
  `derive_head_intent`, `this_term_for_assertion`, `body_digest`
- `analyzer/field.rs` — field-value translation, scalar coercion,
  meta-key detection, unbound-variable collection

`analyzer.rs` becomes the orchestrator: the `analyze()` entry
point + the integration-style test block (kept at root since
tests exercise the public surface end-to-end).

Public surface unchanged: `analyzer::{analyze, AnalyzeError,
Resolver, ResolvedAttribute, ResolvedConcept, NoopResolver,
ResolverError}` re-exported from the module root.

All 69 tests pass; fmt + clippy clean.
