+++ rust/tonk-schema/src/analyzer/field.rs
@@ -1,0 +1,163 @@
+//! Field-value translation — turns parsed [`FieldValue`]s into
+//! the `Term<Any>` slots the engine consumes, plus a few small
+//! utilities (scalar coercion, claim-attribute validation,
+//! reserved meta-field detection, unbound-variable collection).
+
+use std::collections::HashSet;
+
+use dialog_artifacts::{Entity, Value};
+use dialog_query::{Term, attribute::The as AttributeThe};
+use tonk_notation::{Field, FieldValue, Scalar};
+
+use super::error::AnalyzeError;
+use super::resolver::Resolver;
+use super::scope::Scope;
+use crate::transact::{Analysis, Application};
+
+/// Reserved body field names that don't correspond to schema
+/// fields: `this:` (entity selection), `..:` (rest-of-attributes
+/// retraction marker).
+pub(crate) fn is_meta_field(name: &str) -> bool {
+    matches!(name, "this" | "..")
+}
+
+/// Look up the value side of `name` among `fields`, returning
+/// `None` when no field by that name appears.
+pub(crate) fn user_field<'a>(fields: &'a [Field], name: &str) -> Option<&'a FieldValue> {
+    fields.iter().find(|f| f.name == name).map(|f| &f.value)
+}
+
+/// Translate a parsed [`FieldValue`] into the `Term<Any>` slot it
+/// belongs in. Bare symbols resolve at analysis time (against
+/// in-doc declarations first, then the branch's name table);
+/// variables resolve against `analysis.variables` if known,
+/// otherwise stay as `Term::Variable` so planning can substitute
+/// them later; literals become `Term::Constant`; blanks become
+/// `Term::blank()`.
+pub(crate) async fn field_value_to_term<R: Resolver>(
+    field_name: &str,
+    value: &FieldValue,
+    scope: &Scope<'_, R>,
+    analysis: &Analysis,
+) -> Result<Term<dialog_query::Any>, AnalyzeError> {
+    Ok(match value {
+        FieldValue::Literal(scalar) => Term::Constant(scalar_to_value(scalar)?),
+        FieldValue::Variable(name) => {
+            // If this variable was derived in Phase 1, substitute
+            // the entity now; otherwise leave it as a variable
+            // that planning will bind from query results.
+            if let Some(entity) = analysis.variables.get(name) {
+                Term::Constant(Value::Entity(entity.clone()))
+            } else {
+                Term::<dialog_query::Any>::var(name)
+            }
+        }
+        FieldValue::Symbol(name) => {
+            // Bare lowercase symbol — name-table lookup. Same
+            // resolution order as the old `.bookmark` form:
+            //   1. Doc-local declarations (head anchor from the
+            //      same document — `concept!: &foo` or
+            //      `attribute!: &foo`).
+            //   2. Doc-local attribute by name.
+            //   3. Branch entity with `dialog.meta/name = name`.
+            if let Some(entity) = scope.lookup_entity(name) {
+                Term::Constant(Value::Entity(entity))
+            } else if let Some(resolved) =
+                scope
+                    .resolve_attribute(name)
+                    .await
+                    .map_err(|e| AnalyzeError::ResolverFailed {
+                        context: format!("symbol {name}"),
+                        reason: e.message,
+                    })?
+            {
+                Term::Constant(Value::Entity(resolved.entity))
+            } else if let Some(entity) = scope.resolve_named_entity(name).await.map_err(|e| {
+                AnalyzeError::ResolverFailed {
+                    context: format!("symbol {name}"),
+                    reason: e.message,
+                }
+            })? {
+                Term::Constant(Value::Entity(entity))
+            } else {
+                return Err(AnalyzeError::UnknownBookmark {
+                    field: field_name.into(),
+                    bookmark: name.clone(),
+                });
+            }
+        }
+        FieldValue::Uri(uri) => {
+            let entity: Entity =
+                uri.parse()
+                    .map_err(|e: dialog_artifacts::DialogArtifactsError| {
+                        AnalyzeError::InvalidSubjectUri {
+                            subject: uri.clone(),
+                            reason: e.to_string(),
+                        }
+                    })?;
+            Term::Constant(Value::Entity(entity))
+        }
+        FieldValue::Blank => Term::<dialog_query::Any>::blank(),
+        FieldValue::Nested(_) => {
+            return Err(AnalyzeError::UnsupportedFieldValue {
+                field: field_name.into(),
+                form: "nested mapping (only `concept!`'s `with:` accepts a nested map)",
+            });
+        }
+    })
+}
+
+pub(crate) fn scalar_to_value(scalar: &Scalar) -> Result<Value, AnalyzeError> {
+    Ok(match scalar {
+        Scalar::String(s) => Value::String(s.clone()),
+        Scalar::Boolean(b) => Value::Boolean(*b),
+        Scalar::Integer(i) => Value::SignedInt(*i),
+        Scalar::UnsignedInteger(u) => Value::UnsignedInt(*u),
+        Scalar::Float(f) => Value::Float(*f),
+        Scalar::Null => {
+            return Err(AnalyzeError::UnsupportedFieldValue {
+                field: "<scalar>".into(),
+                form: "null literal",
+            });
+        }
+    })
+}
+
+pub(crate) fn scalar_to_string(scalar: &Scalar) -> Result<String, AnalyzeError> {
+    Ok(match scalar {
+        Scalar::String(s) => s.clone(),
+        Scalar::Boolean(b) => b.to_string(),
+        Scalar::Integer(i) => i.to_string(),
+        Scalar::UnsignedInteger(u) => u.to_string(),
+        Scalar::Float(f) => f.to_string(),
+        Scalar::Null => {
+            return Err(AnalyzeError::UnsupportedFieldValue {
+                field: "<scalar>".into(),
+                form: "null literal",
+            });
+        }
+    })
+}
+
+pub(crate) fn validate_claim_attribute(domain: &str, field: &str) -> Result<(), AnalyzeError> {
+    let uri = format!("{domain}/{field}");
+    uri.parse::<AttributeThe>()
+        .map(|_| ())
+        .map_err(|e| AnalyzeError::InvalidClaimAttribute {
+            domain: domain.to_owned(),
+            field: field.to_owned(),
+            reason: format!("{e}"),
+        })
+}
+
+pub(crate) fn collect_unbound_variables(
+    application: &Application,
+    analysis: &Analysis,
+    out: &mut HashSet<String>,
+) {
+    for name in application.bindings() {
+        if !analysis.variables.contains_key(&name) {
+            out.insert(name);
+        }
+    }
+}
